fix(nie-explorer): l'app ne démarrait plus (panic specta), + frameless réel et éditeur de propriétés
L'app ne se lançait plus du tout en dev : le portage nie-trace live avait introduit des DTO à
champs u64/i64, or specta refuse d'exporter les types BigInt vers TypeScript et ce refus fait
paniquer le thread d'export des bindings avant même la création de la fenêtre. Tous les champs
concernés passent en f64 (valeurs bornées très en dessous de 2^53, conversion sans perte,
contrairement à un `as u32` qui tronquerait).

Frameless réel :
- `setAsWindowMenu()` retiré — une barre de menu native suppose une frame native, d'où le retour
  de la bordure, de la légende et des boutons système malgré `decorations: false`.
- `window_vibrancy::apply_mica` retiré pour la même raison (extension de la frame DWM dans la zone
  client) ; `set_titlebar_theme` ne fait plus que `DWMWA_USE_IMMERSIVE_DARK_MODE`.
- Permissions `core:window:*` accordées (toutes celles du schéma) : sans elles, réduire/agrandir/
  fermer et le déplacement de la fenêtre étaient refusés silencieusement.
- 8 poignées `startResizeDragging` : sans cadre natif, la fenêtre n'était plus redimensionnable.

UI portée sur les tokens exacts de var/spaceui (theme.css + 6 variantes), layout ShellLayout de
spacedrive (barre latérale unique, barre supérieure absolue, inspecteur flottant). Les deux barres
latérales concurrentes sont fusionnées en une seule, comme SpacesSidebar amont.

Bugs corrigés : palette de commandes qui plantait l'app (CommandDialog n'enveloppait pas ses
enfants dans <Command>, contexte cmdk manquant) ; boutons imbriqués dans ModsView (Switch et champ
de renommage inutilisables) ; « Ajouter à un mod… » jamais passé au menu contextuel ; menu
contextuel absent sur les entrées de barre latérale ; recherche sans garde anti-race ; <select>
HTML brut de DetailPane.

Nouveau : éditeur de propriétés (fichiers liés, .cfg.bin éditables, fonctions/adresses de
niers.sqlite pour un même code interne), journal durable des opérations (table jobs), et 4 modules
nie-data de plus (boutiques, stades, passifs, tactiques spéciales).

+++ apps/nie-explorer/src-tauri/src/re_trace.rs
@@ -44,7 +44,14 @@ pub fn re_trace_find_process() -> Option<ReTraceProcessDto> {
 pub struct ReTraceRegionDto {
     start: String,
     end: String,
-    size: u64,
+    /// Taille de la plage en octets. `f64` et **pas** `u64` : `specta` refuse d'exporter les types
+    /// « BigInt » (`u64`/`usize`/`i64`/…) vers TypeScript pour éviter une perte de précision
+    /// silencieuse — et le refus est FATAL (l'export panique au démarrage en debug, ce qui
+    /// empêchait l'app de se lancer du tout). Une plage d'un module Windows x64 est bornée par
+    /// l'espace d'adressage utilisateur (2⁴⁷), très en dessous des 2⁵³ entiers exactement
+    /// représentables en `f64` : la conversion est donc SANS perte ici, contrairement à un
+    /// `as u32` qui tronquerait pour de vrai.
+    size: f64,
     perms: String,
     path: String,
 }
@@ -53,7 +60,7 @@ fn to_region_dto(m: &nie_trace::MapEntry) -> ReTraceRegionDto {
     ReTraceRegionDto {
         start: format!("0x{:x}", m.start),
         end: format!("0x{:x}", m.end),
-        size: m.size(),
+        size: m.size() as f64,
         perms: m.perms.clone(),
         path: m.path.clone(),
     }
@@ -95,7 +102,9 @@ fn parse_addr(s: &str) -> Result<u64, String> {
 #[derive(Serialize, specta::Type)]
 pub struct ReTraceDumpStatsDto {
     regions: u32,
-    bytes: u64,
+    /// Octets écrits. `f64` pour la même raison que [`ReTraceRegionDto::size`] — total borné par
+    /// la taille du module dumpé, donc exactement représentable.
+    bytes: f64,
     /// Dossier de sortie réel (sous `BaseDirectory::AppData`, jamais dans le dossier du jeu —
     /// même convention que le workspace de mods, cf. `lib.rs`).
     out_dir: String,
@@ -121,7 +130,7 @@ pub fn re_trace_dump_module(pid: i32, app: tauri::AppHandle) -> Result<ReTraceDu
     let stats = nie_trace::dump_regions(pid, &regions, &out_dir).map_err(|e| e.to_string())?;
     Ok(ReTraceDumpStatsDto {
         regions: stats.regions as u32,
-        bytes: stats.bytes,
+        bytes: stats.bytes as f64,
         out_dir: out_dir.display().to_string(),
     })
 }
